move AppendVecScan to new file
